+++ azbankgateways/banks/__init__.py
@@ -6,5 +6,4 @@ from .mellat import Mellat  # noqa
 from .payV1 import PayV1  # noqa
 from .sep import SEP  # noqa
 from .zarinpal import Zarinpal  # noqa
-from .zarinpalV4 import ZarinpalV4  # noqa
 from .zibal import Zibal  # noqa
